:bat: Changing alignment of dynamic memory chunks and updating to beta version 0.2.017
Switching to serializing dynamic memory with minimum 8 byte alignment instead of 16. This is for a few reasons:

- Most data types don't have 16 byte alignment, so we'd be wasting bytes and generating a bigger serialized blob than we'd need to.

- There's not a clean way to alloc the result buffer with an initial alignment of 16 (or more) without having to return an extra pointer to the user to free. But most allocators do align to 8 internally.

- The thing that requires more than 8 byte alignment most often (at least for my use case) would be data you want to do SIMD aligned loads from, so 128, 256 or 512-bit aligned loads into SIMD registers. But apparently unaligned and aligned loads are equally as fast on current hardware so this shouldn't be that much of a concern.

Ideally we'd like to align the dynamic memory elements to the actual alignment their type has, but we don't have an alignment_of(T) or an alignment field in the Type_Info so no luck there. The language in general doesn't expose much in terms of alignment at the time of writing.

Also ideally we'd love to have allocators support alignment natively so we don't have to keep two pointers, one for the data and another one only for freeing later.

Finally, had to remove some usages of `#must` since it has been unfortunately deprecated.